Initial implementation of the v2 client based on RestSharp
Some notes:

- the old client cannot continue because it uses unmaintained OAuth1.0a lib

- the old client has no support for OAuth2

- the old client returns the resource <T> directly (just like Lyo)
This is a big problem because (a) it assumes a happy path, always getting 200 OK, and (b) not really giving access to all the headers or letting the OslcClient to be reused. This time we return an IOslcResponse that wraps the original reponse. Using a property for the <T> resource also allows us to be lazy about it.

In the future, this allows us to introduce property types for POCOs like OslcLink<T>, where that class would have a method .Dereference(Client c) and that would give you an IOslcResponse, allowing to fully crawl.

Two big questions left to figure out are auth and serialisation.

For auth, RestSharp has a neatly pluggable  Authenticator concept. Unfortunately, it is not a perfect abstraction as Oauth often operates over multiple HTTP requests and requires some interaction. Thus, for the OSLC client we need a bit different abstraction to capture the authn beyond one request.

For serialisation, I am not thrilled to see how much code is in the RdfXmlMediaTypeFormatter. I would prefer the logic to be much simpler and to only focus on supporting multiple RDF formats. Historically, serialization support for OSLC SDKs tried to some of the following:

- OSLC paging support - should be handled via IOslcResponse
- OSLC query response detection - should be handled explicitly by the client making a query and expecting a query response, instead of always checking for it
- RDF/XML-ABBREV intercept - this one may be a bit tricky to get rid of due to Jazz
- some very clever checks if this class can read/write RDF for some class, mostly due to the likely presence of the RDF-XML/ABBREV reader with an overlapping MIME type. , We need to investigate if ASP.NET or RestSharp allow simple means to select the deserialiser and use that instead.